Use Double instead of Int for filters and predicates

+++ src/main/scala/org/zalando/spearheads/innkeeper/api/JsonDtos.scala
@@ -89,9 +89,9 @@ object MatcherType {
   val REGEX = "REGEX"
 }
 
-case class Predicate(name: String, args: Seq[Either[Int, String]])
+case class Predicate(name: String, args: Seq[Either[Double, String]])
 
-case class Filter(name: String, args: Seq[Either[Int, String]])
+case class Filter(name: String, args: Seq[Either[Double, String]])
 
 sealed trait HeaderMatcher {
   def name(): String
